requirements.txt
flask==2.3.3
boto3==1.28.40
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
gunicorn==21.2.0
flask-wtf==1.1.1
pillow==10.0.0
numpy==1.24.0
pandas==1.5.3
plotly==5.16.1
pytest==7.4.0
flask-login==0.6.2
Flask-Cors==3.0.10
